Refactoring of notebooks
Moved notebooks test under modules
