+++ tests/api/test_stack.py
@@ -5,7 +5,6 @@ def test_update_stack(
     pipeline,
     env_fixture,
     dataset_fixture,
-    sgm_notebook,
     sgm_studio,
     cluster,
 ):
@@ -25,11 +24,6 @@ def test_update_stack(
     assert (
         response.data.updateStack.targetUri == sgm_studio.sagemakerStudioUserProfileUri
     )
-
-    response = update_stack_query(
-        client, sgm_notebook.notebookUri, 'notebook', group.name
-    )
-    assert response.data.updateStack.targetUri == sgm_notebook.notebookUri
 
     response = update_stack_query(client, cluster.clusterUri, 'redshift', group.name)
     assert response.data.updateStack.targetUri == cluster.clusterUri
